+++ packages/api/lambdas/sf-event-sqs-to-db-records/index.js
@@ -35,10 +35,12 @@ const {
 const {
   shouldWriteExecutionToPostgres,
   writeExecution,
+  writeExecutionToDynamoAndES,
 } = require('./write-execution');
 
 const {
   writePdr,
+  writePdrToDynamoAndEs,
 } = require('./write-pdr');
 
 const {
@@ -52,8 +54,14 @@ const writeRecordsToDynamoDb = async ({
   pdrModel = new Pdr(),
 }) => {
   const results = await Promise.allSettled([
-    executionModel.storeExecutionFromCumulusMessage(cumulusMessage),
-    pdrModel.storePdrFromCumulusMessage(cumulusMessage),
+    writePdrToDynamoAndEs({
+      cumulusMessage,
+      pdrModel,
+    }),
+    writeExecutionToDynamoAndES({
+      cumulusMessage,
+      executionModel,
+    }),
     granuleModel.storeGranulesFromCumulusMessage(cumulusMessage),
   ]);
   const failures = results.filter((result) => result.status === 'rejected');
@@ -88,7 +96,7 @@ const writeRecords = async ({
   pdrModel,
 }) => {
   if (!isPostRDSDeploymentExecution(cumulusMessage)) {
-    log.info('Message is not for a post-RDS deployment execution. Writes will only be performed to DynamoDB and not RDS');
+    log.info('Message is not for a post-RDS deployment execution. Writes will only be performed to DynamoDB/Elasticsearch and not RDS');
     return writeRecordsToDynamoDb({
       cumulusMessage,
       granuleModel,